+++ tests/test_core.py
@@ -2,6 +2,24 @@ import pyarrow as pa
 import pyarrow.compute as pc
 import pytest
 from graphique.core import Agg, ListChunk, Column as C, Table as T
+from graphique.scalars import parse_duration, duration_isoformat
+
+
+def test_duration():
+    assert duration_isoformat(parse_duration('P1Y1M1DT1H1M1.1S')) == 'P13M1DT3661.100000000S'
+    assert duration_isoformat(parse_duration('P1M1DT1H1M1.1S')) == 'P1M1DT3661.100000000S'
+    assert duration_isoformat(parse_duration('P1DT1H1M1.1S')) == 'P1DT3661.100000S'
+    assert duration_isoformat(parse_duration('PT1H1M1.1S')) == 'PT3661.100000S'
+    assert duration_isoformat(parse_duration('PT1M1.1S')) == 'PT61.100000S'
+    assert duration_isoformat(parse_duration('PT1.1S')) == 'PT1.100000S'
+    assert duration_isoformat(parse_duration('PT1S')) == 'PT1S'
+    assert duration_isoformat(parse_duration('P0D')) == 'PT0S'
+    assert duration_isoformat(parse_duration('PT0S')) == 'PT0S'
+    assert duration_isoformat(parse_duration('P-1DT-1H')) == 'P-2DT82800S'
+    with pytest.raises(ValueError):
+        duration_isoformat(parse_duration('T1H'))
+    with pytest.raises(ValueError):
+        duration_isoformat(parse_duration('P1H'))
 
 
 def test_dictionary(table):