+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 * Pyarrow >=16 required
 * `group` optimized for datasets
+* `Duration` scalar
+
+### Removed
+* `Interval` type
 
 ## [1.5](https://pypi.org/project/graphique/1.5/) - 2024-01-24
 ### Changed